zeenode/requirements.txt
flask
discord.py
colorama
aiohttp
requests
pyfiglet
datetime
urllib3
